# jaboti_backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/pessoas": {
      "get": {
        "operationId": "PessoasController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Busca parcial em name, email, phone",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipo",
            "required": false,
            "in": "query",
            "description": "Filtra por tipo (CLIENTE/USUARIO)",
            "schema": {
              "enum": [
                "CLIENTE",
                "USUARIO"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PessoaResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "pessoas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "PessoasController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePessoaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PessoaResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pessoas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pessoas/{id}": {
      "get": {
        "operationId": "PessoasController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PessoaResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pessoas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PessoasController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePessoaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PessoaResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pessoas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PessoasController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "pessoas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pessoas/change-password": {
      "post": {
        "operationId": "PessoasController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "pessoas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Jaboti API",
    "description": "API multiempresa de atendimento Jaboti",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "PessoaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "chatName": {
            "type": "object"
          },
          "phone": {
            "type": "object"
          },
          "email": {
            "type": "object"
          },
          "type": {
            "type": "string",
            "enum": [
              "CLIENTE",
              "USUARIO"
            ]
          },
          "role": {
            "type": "string",
            "enum": [
              "OPERATOR",
              "SUPERVISOR",
              "ADMIN"
            ]
          },
          "photoUrl": {
            "type": "object"
          },
          "active": {
            "type": "boolean"
          },
          "online": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "role",
          "active",
          "online",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreatePessoaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "João da Silva"
          },
          "phone": {
            "type": "string",
            "example": "+5511999999999",
            "minLength": 5,
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "example": "joao@empresa.com"
          },
          "chatName": {
            "type": "string",
            "description": "Nome curto para chat",
            "example": "João S."
          },
          "type": {
            "type": "string",
            "enum": [
              "CLIENTE",
              "USUARIO"
            ],
            "example": "CLIENTE"
          },
          "password": {
            "type": "string",
            "description": "Senha inicial (apenas para USUARIO). Se omitida: \"changeme\"",
            "minLength": 6,
            "example": "SenhaForte123"
          },
          "photoUrl": {
            "type": "string",
            "description": "URL da foto / avatar",
            "example": "https://cdn.exemplo.com/avatars/joao.png"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdatePessoaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "João Atualizado"
          },
          "phone": {
            "type": "string",
            "example": "+5511888887777",
            "minLength": 5,
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "example": "joao.novo@empresa.com"
          },
          "chatName": {
            "type": "string",
            "example": "João N."
          },
          "type": {
            "type": "string",
            "enum": [
              "CLIENTE",
              "USUARIO"
            ]
          },
          "password": {
            "type": "string",
            "description": "Nova senha (USUARIO).",
            "minLength": 6,
            "example": "NovaSenhaSegura456"
          },
          "photoUrl": {
            "type": "string",
            "description": "URL da foto / avatar",
            "example": "https://cdn.exemplo.com/avatars/joao2.png"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "senhaAtual": {
            "type": "string",
            "minLength": 6,
            "example": "SenhaAtual123"
          },
          "novaSenha": {
            "type": "string",
            "minLength": 6,
            "example": "NovaSenhaSegura456"
          }
        },
        "required": [
          "senhaAtual",
          "novaSenha"
        ]
      },
...
    }
  }
}